Curve Fitting by using numerical methods (Python)
Curve fitting for given data using Least Square Regression method. Both parabolic and linear function is used to fit the data. At the end, graphical representation is used for both two types of functions in order to compare which function fits the data better. It is clearly seen that for this dataset and in these conditions, linear function is better choice.